upgraded structure

+++ Maths.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <math.h>
 
-static const float PI = atan(1) * 4.f;
+static const double PI = atan(1) * 4.f;
